requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
Django==1.11.29
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.31.3
MarkupSafe==2.1.5
msgpack==1.0.8
packaging==24.1
pipdeptree==2.23.1
psutil==6.0.0
pycparser==2.22
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==26.2.0
requests==2.32.3
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
zipp==3.20.0
zope.event==5.0
zope.interface==7.0.1